third_party/WebKit/LayoutTests/media/media-controller-media-fragment-uri-initial-playback-position-expected.txt

 
EXPECTED (video.currentTime == '2') OK
EXPECTED (video2.currentTime == '0') OK
EXPECTED (video.currentTime == '2') OK
EXPECTED (video2.currentTime == '2') OK
END OF TEST
